+++ packages/remote-bindings/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @cloudflare/remote-bindings
 
+## 0.0.32
+
+### Patch Changes
+
+- Updated dependencies [[`2f5781a`](https://github.com/cloudflare/workers-sdk/commit/2f5781a5916ec2bc5ca0a2047c4c28cc9399227c)]:
+  - @cloudflare/deploy-helpers@0.15.0
+
+## 0.0.31
+
+### Patch Changes
+
+- Updated dependencies [[`93ee76e`](https://github.com/cloudflare/workers-sdk/commit/93ee76e6eb99cbb1d42abb4f472eb00d61c0d7a1), [`59267fc`](https://github.com/cloudflare/workers-sdk/commit/59267fc79d1f7925a15369ca0125290df2404bfb), [`6906bf0`](https://github.com/cloudflare/workers-sdk/commit/6906bf06d9eb1045605c71c345d37e7c300a5bbc)]:
+  - @cloudflare/workers-auth@0.9.0
+  - @cloudflare/workers-utils@0.41.1
+  - @cloudflare/deploy-helpers@0.14.2
+  - @cloudflare/cli-shared-helpers@0.1.36
+  - miniflare@5.20260921.0-alpha
+
 ## 0.0.30
 
 ### Patch Changes